backend/requirements.txt
flask==2.0.1
pymongo==4.0.1
boto3==1.24.0
apscheduler==3.9.1
python-dotenv==0.20.0
flask-cors==3.0.10
flask-limiter==3.3.1
gunicorn==20.1.0
pytz==2022.1
requests==2.28.0
werkzeug==2.0.2
itsdangerous==2.0.1
jinja2==3.0.3
markupsafe==2.0.1
click==8.0.3
six==1.16.0
tzlocal==4.2
setuptools==65.5.1
wheel==0.38.4